Apply single responsibility on LocationController & Generate stubs
- LocationController cleaned, validation rules moved to LocationRequest
- Uploading images reused upload_base64_image()
- Stubs generated
- stubs/request.stub updated to return api response on failure

+++ app/Http/Controllers/Api/Admin/Course/LocationController.php
@@ -5,9 +5,10 @@ namespace App\Http\Controllers\Api\Admin\Course;
 use Illuminate\Http\Request;
 use App\Models\Course\Location;
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Validator;
-use App\Http\Resources\Admin\Course\NamingResource;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
+use App\Http\Requests\Admin\Course\LocationRequest;
+use App\Http\Resources\Admin\Course\NamingResource;
 
 class LocationController extends Controller
 {
@@ -22,60 +23,30 @@ class LocationController extends Controller
         $locations = Location::filter($request)->sortData($request)->offset($request->perPage * $request->page)->paginate($request->perPage);
 
         return response()->json([
-          'total'   => Location::all()->count(),
-          'namings' => NamingResource::collection($locations)
+            'total'   => Location::all()->count(),
+            'namings' => NamingResource::collection($locations)
         ]);
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  LocationRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(LocationRequest $request)
     {
-        // Validation
-        $validator = Validator::make($request->all(), [
-          'name_ar' => 'required|min:3|unique:locations',
-          'name_en' => 'nullable|min:3',
-          'description_ar' => 'nullable|min:3',
-          'description_en' => 'nullable|min:3'
-        ]);
-
-        if ($validator->fails()) {
-          return response()->json($validator->errors(), 400);
-        }
-
         // Upload Image
         if ($request->image) {
-          $base64Image  = explode(";base64,", $request->image);
-          $explodeImage = explode("image/", $base64Image[0]);
-          $imageLocation    = $explodeImage[1];
-          $image_base64 = base64_decode($base64Image[1]);
-          $imageName    = uniqid() . '.'.$imageLocation;
-          Storage::disk('public')->put("courses/namings/images/{$imageName}", $image_base64);
-          $request->merge(['image' => $imageName]);
+            $name = upload_base64_image($request->image, "courses/namings/images");
+            $request->merge(['image' => $name]);
         }
-
         // Store in database
         $location = Location::create($request->all());
-
         return response()->json([
-          'message' => __('messages.successful_create'),
-          'subscriber' => new NamingResource($location)
+            'message' => __('messages.successful_create'),
+            'subscriber' => new NamingResource($location)
         ], 200);
-
     }
 
     /**
@@ -86,57 +57,41 @@ class LocationController extends Controller
      */
     public function show(Location $location)
     {
-      return new NamingResource($location);
+        return new NamingResource($location);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  LocationRequest  $request
      * @param  Location  $location
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Location $location)
+    public function update(LocationRequest $request, Location $location)
     {
-        // Validation
-        $validator = Validator::make($request->all(), [
-          'name_ar' => 'required|min:3|unique:locations,name_ar,' . $location->id,
-          'name_en' => 'nullable|min:3',
-          'description_ar' => 'nullable|min:3',
-          'description_en' => 'nullable|min:3'
-        ]);
 
-        if ($validator->fails()) {
-          return response()->json($validator->errors(), 400);
+        // Delete the previous image
+        if (!$request->image && $location->image) {
+            Storage::disk('public')->delete("courses/namings/images/{$location->image}");
         }
 
         // Upload Image
         if ($request->image) {
-          if (str_starts_with($request->image, 'data:image')) {
-            $base64Image  = explode(";base64,", $request->image);
-            $explodeImage = explode("image/", $base64Image[0]);
-            $imageLocation    = $explodeImage[1];
-            $image_base64 = base64_decode($base64Image[1]);
-            $imageName    = uniqid() . '.'.$imageLocation;
-            // Delete the previous image
-            Storage::disk('public')->delete("courses/namings/images/{$location->image}");
-            // Save the new image
-            Storage::disk('public')->put("courses/namings/images/{$imageName}", $image_base64);
-            $request->merge(['image' => $imageName]);
-          } else {
-            $request->merge(['image' => $location->image]);
-          }
-        } else if($location->image) {
-          // Delete the previous image
-          Storage::disk('public')->delete("courses/namings/images/{$location->image}");
+            if (str_starts_with($request->image, 'data:image')) {
+                // Delete the previous image
+                Storage::disk('public')->delete("courses/namings/images/{$location->image}");
+                // Save the new image
+                $name = upload_base64_image($request->image, "courses/namings/images");
+                $request->merge(['image' => $name]);
+            }
         }
 
         // Store in database
         $location->update($request->all());
 
         return response()->json([
-          'message' => __('messages.successful_update'),
-          'subscriber' => new NamingResource($location)
+            'message' => __('messages.successful_update'),
+            'subscriber' => new NamingResource($location)
         ], 200);
     }
 
@@ -151,7 +106,7 @@ class LocationController extends Controller
         $location->status = !$location->status;
         $location->save();
         return response()->json([
-          'message' => __('messages.successful_update')
+            'message' => __('messages.successful_update')
         ], 200);
     }
 
@@ -170,8 +125,7 @@ class LocationController extends Controller
         // Delete database record
         $location->delete();
         return response()->json([
-          'message' => __('messages.successful_delete')
+            'message' => __('messages.successful_delete')
         ], 200);
     }
-  
 }
